Use file.rename() on Windows with staged install to move package to its final location.
git-svn-id: https://svn.r-project.org/R/trunk@76171 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/R/install.R
@@ -1599,9 +1599,13 @@ if(FALSE) {
 
         if (pkg_staged_install) {
             if (WINDOWS) {
-                file.copy(instdir, dirname(final_instdir), recursive = TRUE,
-                          copy.date = TRUE)
-                unlink(instdir, recursive = TRUE)
+                unlink(final_instdir, recursive = TRUE) # needed for file.rename
+                if (!file.rename(instdir, final_instdir)) {
+                    message("WARNING: moving package to final location failed, copying instead")
+                    file.copy(instdir, dirname(final_instdir), recursive = TRUE,
+                              copy.date = TRUE)
+                    unlink(instdir, recursive = TRUE)
+                }
             } else {
                 patch_rpaths()
